+++ fluentlenium-junit-jupiter/src/it/junit-jupiter/junit-jupiter-default/src/test/java/org/fluentlenium/it/DefaultTest.java
@@ -1,0 +1,20 @@
+package org.fluentlenium.it;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.fluentlenium.adapter.junit.jupiter.FluentTest;
+import org.junit.jupiter.api.Test;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.htmlunit.HtmlUnitDriver;
+
+public class DefaultTest extends FluentTest {
+    @Override
+    public WebDriver newWebDriver() {
+        return new HtmlUnitDriver(true);
+    }
+
+    @Test
+    public void test() {
+        assertTrue(true);
+    }
+}
