+++ pom.xml
@@ -14,7 +14,7 @@
         <module>fluentlenium-core</module>
         <module>fluentlenium-assertj</module>
         <module>fluentlenium-junit</module>
-        <module>fluentlenium-junit5</module>
+        <module>fluentlenium-junit-jupiter</module>
         <module>fluentlenium-testng</module>
         <module>fluentlenium-cucumber</module>
         <module>fluentlenium-spock</module>
